+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## dev
+* Added Vue components regex
 * Fixed [Alphabetical sorting doesn't work](https://github.com/7ute/symbols-list/issues/68)
 
 ## 2.5.2 - Multiline JS, SASS, Twig and Blade